Refactor OrderInfo into OrderTransactionInfo
