+++ src/lib/Order.ts
@@ -3,9 +3,8 @@ import Web3 from 'web3';
 
 // Types
 import { Provider } from '@0xproject/types';
-import { ECSignature, MarketContract, Order, OrderLib, SignedOrder } from '@marketprotocol/types';
+import { ECSignature, Order, OrderLib, SignedOrder } from '@marketprotocol/types';
 
-import { MarketError } from '../types';
 import { Utils } from './Utils';
 import { assert } from '../assert';
 
@@ -166,235 +165,4 @@ export async function signOrderHashAsync(
   const web3: Web3 = new Web3();
   web3.setProvider(provider);
   return Utils.signMessage(web3, signerAddress, orderHash);
-}
-
-/**
- * Parameters to initialize OrderInfo
- *
- */
-export interface OrderInfoParams {
-  txHash: string;
-  blockNumber?: number;
-}
-
-/**
- * OrderInfo fetches and hold all the necessary information about an order
- * that has been posted through a market contract.
- *
- */
-export class OrderInfo {
-  // region Members
-  // *****************************************************************
-  // ****                     Members                             ****
-  // *****************************************************************
-  static readonly ORDER_EXPIRED_CODE = '0';
-  static readonly ORDER_DEAD_CODE = '1';
-
-  readonly txHash: string;
-
-  private readonly _marketContract: MarketContract;
-  private readonly _order: Order;
-  private _fromBlockNumber: number = 0;
-  private _toBlockNumber?: number;
-
-  // endregion // Members
-
-  // region Constructors
-  // *****************************************************************
-  // ****                     Constructors                        ****
-  // *****************************************************************
-
-  private constructor(marketContract: MarketContract, order: Order, txHash: string) {
-    this._marketContract = marketContract;
-    this._order = order;
-    this.txHash = txHash;
-  }
-
-  // endregion // Constructors
-  // region Properties
-  // *****************************************************************
-  // ****                     Properties                          ****
-  // *****************************************************************
-
-  /**
-   * Factory function to create all relevant order information
-   *
-   * @param {MarketContract} marketContract MarketContract for order
-   * @param {Order} order Order with info to be tracked
-   * @param {OrderInfoParams} txParams transaction parameters
-   */
-  static create(
-    marketContract: MarketContract,
-    order: Order,
-    txParams: OrderInfoParams
-  ): OrderInfo {
-    const orderInfo = new OrderInfo(marketContract, order, txParams.txHash);
-    orderInfo._fromBlockNumber = txParams.blockNumber ? txParams.blockNumber : 0;
-    orderInfo._toBlockNumber = txParams.blockNumber;
-    return orderInfo;
-  }
-
-  // endregion //Properties
-
-  // region Public Methods
-  // *****************************************************************
-  // ****                     Public Methods                      ****
-  // *****************************************************************
-
-  /**
-   * Fetches filled quantity for this order.
-   *
-   *
-   */
-  get filledQty(): Promise<BigNumber | number> {
-    return (async () => {
-      let stopEventWatcher = async () => {
-        return;
-      }; // pre-initialized to prevent NPE checks :)
-
-      try {
-        const fillQty = await Promise.race([
-          this._fetchOrWatchForFilledQty(stopEventWatcher),
-          this._watchForError()
-        ]);
-        return fillQty;
-      } catch (err) {
-        await stopEventWatcher(); // cleanup, prevent further watching
-        return Promise.reject(err);
-      }
-    })();
-  }
-
-  /**
-   * Fetches cancelled quantity for this order
-   *
-   */
-  get cancelledQty(): Promise<BigNumber | number> {
-    return (async () => {
-      let stopEventWatcher = async () => {
-        return;
-      }; // pre-initialized to prevent NPE checks :)
-
-      try {
-        const cancelledQty = await Promise.race([
-          this._fetchOrWatchForCancelledQty(stopEventWatcher),
-          this._watchForError()
-        ]);
-        return cancelledQty;
-      } catch (err) {
-        await stopEventWatcher(); // cleanup, prevent further watching
-        return Promise.reject(err);
-      }
-    })();
-  }
-
-  // endregion // Public Methods
-  // region Private Methods
-  // *****************************************************************
-  // ****                     Private Methods                     ****
-  // *****************************************************************
-
-  /**
-   * Tries to fetch filledQty for this Order.
-   * If nothing found, watches for when it is found.
-   *
-   * @param {() => Promise<void>} stopWatcher stop watching for filled Qty
-   */
-  private _fetchOrWatchForFilledQty(stopWatcher: () => Promise<void>): Promise<BigNumber | number> {
-    return new Promise<BigNumber | number>(async (resolve, reject) => {
-      const watchFilter = { fromBlock: this._fromBlockNumber, toBlock: this._toBlockNumber };
-      const orderEvent = this._marketContract.OrderFilledEvent({ maker: this._order.maker });
-
-      // try fetching event
-      const eventLogs = await orderEvent.get(watchFilter);
-      let foundEvent = eventLogs.find(eventLog => eventLog.transactionHash === this.txHash);
-      if (foundEvent) {
-        resolve(foundEvent.args.filledQty);
-        return;
-      }
-
-      // watch for event
-      stopWatcher = orderEvent.watch(watchFilter, (err, eventLog) => {
-        if (err) {
-          console.log(err);
-        }
-
-        if (eventLog.transactionHash === this.txHash) {
-          stopWatcher()
-            .then(function() {
-              return resolve(eventLog.args.filledQty);
-            })
-            .catch(reject);
-        }
-      });
-    });
-  }
-
-  /**
-   * Tries to fetch cancelled for this Order.
-   * If nothing found, watches for when it is found.
-   *
-   * @param {() => Promise<void>} stopWatcher stop watching for cancelled Qty
-   */
-  private _fetchOrWatchForCancelledQty(
-    stopWatcher: () => Promise<void>
-  ): Promise<BigNumber | number> {
-    return new Promise<BigNumber | number>(async (resolve, reject) => {
-      const watchFilter = { fromBlock: this._fromBlockNumber, toBlock: this._toBlockNumber };
-      const orderEvent = this._marketContract.OrderCancelledEvent({ maker: this._order.maker });
-
-      // try fetching event
-      const eventLogs = await orderEvent.get(watchFilter);
-      let foundEvent = eventLogs.find(eventLog => eventLog.transactionHash === this.txHash);
-      if (foundEvent) {
-        resolve(foundEvent.args.cancelledQty);
-        return;
-      }
-
-      // watch for event
-      stopWatcher = orderEvent.watch(watchFilter, (err, eventLog) => {
-        if (err) {
-          console.log(err);
-        }
-
-        if (eventLog.transactionHash === this.txHash) {
-          stopWatcher()
-            .then(function() {
-              return resolve(eventLog.args.cancelledQty);
-            })
-            .catch(reject);
-        }
-      });
-    });
-  }
-
-  private _watchForError(): Promise<BigNumber | number> {
-    return new Promise((_, reject) => {
-      const errorEvent = this._marketContract.ErrorEvent({});
-
-      const stopErrorEventWatcher = errorEvent.watch(
-        { fromBlock: this._fromBlockNumber, toBlock: this._toBlockNumber },
-        (err, eventLog) => {
-          if (err) {
-            console.log(err);
-          }
-
-          if (eventLog.transactionHash === this.txHash) {
-            stopErrorEventWatcher()
-              .then(() => {
-                switch (eventLog.args.errorCode.toString()) {
-                  case OrderInfo.ORDER_EXPIRED_CODE:
-                    return reject(new Error(MarketError.OrderExpired));
-                  case OrderInfo.ORDER_DEAD_CODE:
-                    return reject(new Error(MarketError.OrderDead));
-                  default:
-                    return reject(new Error(MarketError.UnknownOrderError));
-                }
-              })
-              .catch(reject);
-          }
-        }
-      );
-    });
-  }
 }
